requirements.txt
Flask==3.1.2
Jinja2==3.1.6
gunicorn==23.0.0
python-dotenv==1.2.1
jsonschema==4.25.1
pytest==8.4.2
black==25.11.0
ruff==0.14.0
